+++ app/llm-evaluation-dashboard/constants.ts
@@ -1,0 +1,14 @@
+export const COLORS: Record<string, string> = {
+  google: "hsl(var(--chart-1))",
+  openai: "hsl(var(--chart-2))",
+  xai: "hsl(var(--chart-3))",
+  googleLight: "hsl(var(--chart-1) / 0.5)",
+  openaiLight: "hsl(var(--chart-2) / 0.5)",
+  xaiLight: "hsl(var(--chart-3) / 0.5)",
+};
+
+export const PROVIDER_LABELS: Record<string, string> = {
+  google: "Google (Gemini)",
+  openai: "OpenAI (GPT)",
+  xai: "xAI (Grok)",
+};
